% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCWaterDischargeAnalysis.R
\docType{package}
\name{BCWaterDischargeAnalysis-package}
\alias{BCWaterDischargeAnalysis}
\alias{BCWaterDischargeAnalysis-package}
\title{Performs analysis on water discharge from streams}
\description{
This is a suite of functions for the analysis of stream discharge data..
}
