% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.Q.percentile.longterm.R
\name{compute.Q.percentile.longterm}
\alias{compute.Q.percentile.longterm}
\title{Compute long term annual (calendar year) percentiles.}
\usage{
compute.Q.percentile.longterm(Station.Code = "XXXXX", Station.Area = NA,
  flow, start.year = 9999, end.year = 0, per.list = c(1, 2, seq(5, 95, 5),
  98, 99), write.stat.csv = TRUE, write.stat.trans.csv = TRUE,
  report.dir = ".", csv.nddigits = 3, na.rm = list(na.rm.global = TRUE),
  debug = FALSE)
}
\arguments{
\item{Station.Code}{A character string. Used to names any *.csv or *.pdf files created so be careful
that the Station.Code will be acceptable in a file name.}

\item{Station.Area}{Area of the water basin (square km), used in computing some statistics}

\item{flow}{Dataframe with two variables (1) \code{Date} with the date as a R date class variable and
         (2) \code{Q} a numeric variable with the flow for that day.

         Missing values for \code{Q} can be
         explicitly represented using \code{NA} with the associated date, or implicitly represented
         by excluding the date from the data.frame.

         The date range of \code{flow} can be wider than
         the \code{start.year} and \code{end.year}. This is useful for statistics on the water year
         (starting 1 October of the previous year) when you may wish to include flow data for the latter
         part of the year. Similarly, it may be useful for statistics that use a 3, 7, or 30 day
         rolling average as this will be missing for the first few days of each year.

         A water year runs from 1 October of the previous year to 30 Sept of the current year

         All other variables in the data frame will be ignored. Data does NOT have to be sorted by Date order.}

\item{start.year, end.year}{Starting and ending year for statistics in YYYY format, e.g. start.year=2013.}

\item{per.list}{Percentiles to be computed}

\item{write.stat.csv}{Should a file be created with the computed percentiles?
The file name will be  \code{file.path(report.dir,paste(Station.Code,'-longterm-percentile-stat.csv'))}.}

\item{write.stat.trans.csv}{Should a file be created with the transposed of the percentile report?
The file name will be  \code{file.path(report.dir,paste(Station.Code,'-longterm-percentile-stat-trabs.csv'))}.}

\item{report.dir}{Directory in which *.csv or *.pdf files are stored.}

\item{csv.nddigits}{Number of decimal digits to be rounded to when creating *.csv files.}

\item{na.rm}{List of logical values indicating how to deal with missing values.
\code{na.rm$na.rm.global} control if missing values are to be globally excluded (TRUE) or propogated (FALSE).
If missing values are to be propogated, then any statistic computed over a missing value is also set
to missing. For example, if a flow value was missing on 2014-01-03, then any statistics for January of
2014, JFM of 2014, calendar year 2014, water year 2014, 3 day rolling average of 1, 2, 3 of January 2014
and all statistics that depend on those rolling averages, will also be set to missing.}

\item{debug}{Internal flag used for debugging}
}
\value{
A list with the following elements:
  \item{Q.percentile.stat}{Data frame with the percentiles of \code{flow$Q} by month
        and overall between \code{start.year} and \code{end.year}}
  \item{Q.percentile.stat.trans}{Data frame with the percentiles of \code{flow$Q} transposed.}
  \item{file.stat.csv}{File name of *.csv file with percentile statistics.}
   \item{file.stat.trans.csv}{File name of *.csv file with transposed percentile statistics.}
   \item{na.rm}{Missing value flags.}
   \item{Version}{Version of this function.}
   \item{Date}{Date function was run.}
}
\description{
Computes many percentiles on the \code{flow$Q} variable
   between \code{start.year} and \code{end.year} inclusive.
   It (optionally) saves the resutls in *.csv and *.pdf files.
}
\examples{
\dontrun{
percentile.longterm <- compute.Q.percentile.longterm(
                         Station.Code  ='ABCDE',
                         Station.Area  =2458,
                         flow          =flow,
                         start.year    =1960,
                         end.year      =2015)
}

}
