% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.Q.stat.annual.R
\name{compute.Q.stat.annual}
\alias{compute.Q.stat.annual}
\title{Compute annual (calendar and water year) statistics.}
\usage{
compute.Q.stat.annual(Station.Code = "XXXXX", Station.Area = NA, flow,
  start.year = 9999, end.year = 0, write.cy.stat.csv = TRUE,
  write.wy.stat.csv = TRUE, write.stat.trans.csv = TRUE,
  write.flow.summary.csv = TRUE, write.lowflow.csv = TRUE,
  plot.stat.trend = TRUE, plot.cumdepart = TRUE, report.dir = ".",
  na.rm = list(na.rm.global = FALSE), csv.nddigits = 3, debug = FALSE)
}
\arguments{
\item{Station.Code}{A character string. Used to name any *.csv or *.pdf files created so be careful
that the Station.Code will be acceptable in a file name.}

\item{Station.Area}{Area of the water basin (square km), used in computing some statistics}

\item{flow}{Dataframe with two variables (1) \code{Date} with the date as a R date class variable and
         (2) \code{Q} a numeric variable with the flow for that day.

         Missing values for \code{Q} can be
         explicitly represented using \code{NA} with the associated date, or implicitly represented
         by excluding the date from the data.frame.

         The date range of \code{flow} can be wider than
         the \code{start.year} and \code{end.year}. This is useful for statistics on the water year
         (starting 1 October of the previous year) when you may wish to include flow data for the latter
         part of the year. Similarly, it may be useful for statistics that use a 3, 7, or 30 day
         rolling average as this will be missing for the first few days of each year.

         A water year runs from 1 October of the previous year to 30 Sept of the current year

         All other variables in the data frame will be ignored. Data does NOT have to be sorted by Date order.}

\item{start.year, end.year}{Starting and ending year for statistics in YYYY format, e.g. start.year=2013.}

\item{write.cy.stat.csv}{Should a file be created with the calendar year computed percentiles?
The file name will be  \code{file.path(report.dir,paste(Station.Code,'-annual-cy-summary-stat.csv'))}.}

\item{write.wy.stat.csv}{Should a file be created with the water year computed percentiles?
The file name will be  \code{file.path(report.dir,paste(Station.Code,'-annual-wy-summary-stat.csv'))}.}

\item{write.stat.trans.csv}{Should a file be created with the transposed of the annual statistics
(both calendar and water year)?
The file name will be  \code{file.path(report.dir,paste(Station.Code,'-annual-summary-stat-trans.csv'))}.}

\item{write.flow.summary.csv}{Should a file be created with a flow summary over the years between the
start.year and end.year (inclusive). This summary includes number of days, number of missing values,
mean, median, minimum, maximum, and standard deviation of \code{flow$Q}.
The file name will be \code{file.path(report.dir, paste(Station.Code,"-period-record-summary.csv", sep=""))}.}

\item{write.lowflow.csv}{Should a file be created with the minimum value of \code{flow$Q} and date the
minimum occured.
The file name will be \code{file.path(report.dir,paste(Station.Code,"-lowflow-summary.csv",sep=""))}}

\item{plot.stat.trend}{Should a file be created with plots of the statistics over the years
     between \code{start.year} and \code{end.year}.
The file name will be \code{file.path(report.dir, paste(Station.Code,"-annual-trend.pdf",sep=""))}}

\item{plot.cumdepart}{Should a file be created with plots of the yearly and cumulative departures
from the grand mean between \code{start.year} and \code{end.year}.
The file name will be \code{file.path(report.dir, paste(Station.Code,"-cumulative departure.pdf",sep=""))}}

\item{report.dir}{Directory in which *.csv or *.pdf files are stored.}

\item{na.rm}{List of logical values indicating how to deal with missing values.
\code{na.rm$na.rm.global} control if missing values are to be globally excluded (TRUE) or propogated (FALSE).
If missing values are to be propogated, then any statistic computed over a missing value is also set
to missing. For example, if a flow value was missing on 2014-01-03, then any statistics for January of
2014, JFM of 2014, calendar year 2014, water year 2014, 3 day rolling average of 1, 2, 3 of January 2014
and all statistics that depend on those rolling averages, will also be set to missing.}

\item{csv.nddigits}{Number of decimal digits to be rounded to when creating *.csv files.}

\item{debug}{Internal flag used for debugging}
}
\value{
A list with the following elements:
  \item{Q.flow.summary}{Data frame with flow summary.}
  \item{Q.stat.annual}{Data frame with summary statistics as listed at \code{\link{SummaryStatistics}}.}
  \item{Q.stat.annual.trans}{Data frame with transposed summary statistics as listed at \code{\link{SummaryStatistics}}.}
  \item{dates.missing.flow}{Data framw with dates of missing \code{flow$Q} between
         \code{start.year} and \code{end.year}}
  \item{file.cy.stat.csv}{File name of *.csv file with calendar year summary statistics.}
  \item{file.wy.stat.csv}{File name of *.csv file with water year summary statistics.}
  \item{file.stat.trans.csv}{File name of *.csv file with transposed summary statistics.}
  \item{file.stat.trend.pdf}{File name of *.pdf file with plot of statistics over time.}
  \item{file.cumdepart.pdf}{File name of *.pdf file with plot of yearly and cumulative departures.}
  \item{file.summary.csv}{File name of *.csv file with summary statistics.}
  \item{file.lowflow.csv}{File name of *.csv file with low flow summary statistics.}
  \item{na.rm}{Missing value flags.}
  \item{Version}{Version of this function.}
  \item{Date}{Date function was run.}
}
\description{
Computes many annual (calendar and water year) statistics
and (optionally) saves the resutls in *.csv and *.pdf files.
}
\examples{
\dontrun{
stat.annual <- compute.Q.stat.annual(
                         Station.Code  ='ABCD',
                         Station.Area  =12345,
                         flow          =flow,
                         start.year    =1960,
                         end.year      =2014)
}
}
