% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.Q.stat.longterm.R
\name{compute.Q.stat.longterm}
\alias{compute.Q.stat.longterm}
\title{Compute long-term annual (calendar and water year) summary statistics.}
\usage{
compute.Q.stat.longterm(Station.Code = "XXXXX", flow, start.year = 9999,
  end.year = 0, write.cy.stat.csv = TRUE, write.cy.stat.trans.csv = TRUE,
  write.wy.stat.csv = TRUE, write.wy.stat.trans.csv = TRUE,
  report.dir = ".", csv.nddigits = 3, na.rm = list(na.rm.global = TRUE))
}
\arguments{
\item{Station.Code}{A character string. Used to name any *.csv or *.pdf files created so be careful
that the Station.Code will be acceptable in a file name.}

\item{flow}{Dataframe with two variables (1) \code{Date} with the date as a R date class variable and
         (2) \code{Q} a numeric variable with the flow for that day.

         Missing values for \code{Q} can be
         explicitly represented using \code{NA} with the associated date, or implicitly represented
         by excluding the date from the data.frame.

         The date range of \code{flow} can be wider than
         the \code{start.year} and \code{end.year}. This is useful for statistics on the water year
         (starting 1 October of the previous year) when you may wish to include flow data for the latter
         part of the year. Similarly, it may be useful for statistics that use a 3, 7, or 30 day
         rolling average as this will be missing for the first few days of each year.

         A water year runs from 1 October of the previous year to 30 Sept of the current year

         All other variables in the data frame will be ignored. Data does NOT have to be sorted by Date order.}

\item{start.year, end.year}{Starting and ending year for statistics in YYYY format, e.g. start.year=2013.}

\item{write.cy.stat.csv}{Should a file be created with the computed statistics?
The file name will be  \code{file.path(report.dir,paste(Station.Code,'-longterm-cy-summary-stat.csv'))}.}

\item{write.cy.stat.trans.csv}{Should a file be created with the transposed of the statistics report?
The file name will be  \code{file.path(report.dir,paste(Station.Code,'-longterm-cy-summary-stat-trans.csv'))}.}

\item{write.wy.stat.csv}{Should a file be created with the computed statistics?
The file name will be  \code{file.path(report.dir,paste(Station.Code,'-longterm-wy-summary-stat.csv'))}.}

\item{write.wy.stat.trans.csv}{Should a file be created with the transposed of the statistics report?
The file name will be  \code{file.path(report.dir,paste(Station.Code,'-longterm-wy-summary-stat-trans.csv'))}.}

\item{report.dir}{Directory in which *.csv or *.pdf files are stored.}

\item{csv.nddigits}{Number of decimal digits to be rounded to when creating *.csv files.}

\item{na.rm}{List of logical values indicating how to deal with missing values.
\code{na.rm$na.rm.global} control if missing values are to be globally excluded (TRUE) or propogated (FALSE).
If missing values are to be propogated, then any statistic computed over a missing value is also set
to missing. For example, if a flow value was missing on 2014-01-03, then any statistics for January of
2014, JFM of 2014, calendar year 2014, water year 2014, 3 day rolling average of 1, 2, 3 of January 2014
and all statistics that depend on those rolling averages, will also be set to missing.}
}
\value{
A list with the following elements:
  \item{Q.cy.stat.longterm}{Data frame with the long-term statistics of \code{flow$Q} by month
        and overall between \code{start.year} and \code{end.year}}
  \item{Q.cy.stat.longterm.trans}{Data frame with the long-term statistics of \code{flow$Q} transposed.}
  \item{file.cy.stat.csv}{Object with file name of *.csv file with long term summary statistics.}
   \item{file.cy.stat.trans.csv}{Object with file name of *.csv file with transposed long-term summary statistics.}
   \item{Q.wy.stat.longterm}{Data frame with the long-term statistics of \code{flow$Q} by month
        and overall between \code{start.year} and \code{end.year}}
  \item{Q.wy.stat.longterm.trans}{Data frame with the long-term statistics of \code{flow$Q} transposed.}
  \item{file.wy.stat.csv}{Object with file name of *.csv file with long-term summary statistics.}
   \item{file.wy.stat.trans.csv}{Object with file name of *.csv file with transposed long-term summary statistics.}
   \item{na.rm}{Missing value flags.}
   \item{Version}{Version of this function.}
   \item{Date}{Date function was run.}
}
\description{
Computes long-term summary statistics on \code{flow$Q} variable
   between \code{start.year} and \code{end.year} inclusive.
   It (optionally) saves the resutls in *.csv and *.pdf files.
}
\examples{
\dontrun{
stat.longterm <- compute.Q.stat.longterm(
                         Station.Code  ='ABCDE',
                         flow          =flow,
                         start.year    =1960,
                         end.year      =2015)
}

}
