% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.volume.frequency.analysis.R
\name{compute.volume.frequency.analysis}
\alias{compute.volume.frequency.analysis}
\title{Perform a volume frequency analysis on annual statistics.}
\usage{
compute.volume.frequency.analysis(Station.Code, flow, start.year = 9999,
  end.year = 0, use.water.year = FALSE, roll.avg.days = c(1, 3, 7, 15, 30,
  60, 90), use.log = FALSE, use.max = FALSE,
  prob.plot.position = c("weibull", "median", "hazen"),
  prob.scale.points = c(0.9999, 0.999, 0.99, 0.9, 0.5, 0.2, 0.1, 0.02, 0.01,
  0.001, 1e-04), fit.distr = c("PIII", "weibull"),
  fit.distr.method = ifelse(fit.distr == "PIII", "MOM", "MLE"),
  fit.quantiles = c(0.975, 0.99, 0.98, 0.95, 0.9, 0.8, 0.5, 0.2, 0.1, 0.05,
  0.01), na.rm = list(na.rm.global = TRUE), write.stat.csv = TRUE,
  write.stat.trans.csv = TRUE, write.plotdata.csv = FALSE,
  write.quantiles.csv = TRUE, write.quantiles.trans.csv = TRUE,
  write.frequency.plot = TRUE, write.frequency.plot.suffix = c("pdf",
  "png"), report.dir = ".", csv.nddigits = 3, debug = FALSE)
}
\arguments{
\item{Station.Code}{A character string. Used to names any *.csv or *.pdf files created so be careful
that the Station.Code will be acceptable in a file name.}

\item{flow}{Dataframe with two variables (1) \code{Date} with the date as a R date class variable and
         (2) \code{Q} a numeric variable with the flow for that day.

         Missing values for \code{Q} can be
         explicitly represented using \code{NA} with the associated date, or implicitly represented
         by excluding the date from the data.frame.

         The date range of \code{flow} can be wider than
         the \code{start.year} and \code{end.year}. This is useful for statistics on the water year
         (starting 1 October of the previous year) when you may wish to include flow data for the latter
         part of the year. Similarly, it may be useful for statistics that use a 3, 7, or 30 day
         rolling average as this will be missing for the first few days of each year.

         A water year runs from 1 October of the previous year to 30 Sept of the current year

         All other variables in the data frame will be ignored. Data does NOT have to be sorted by Date order.}

\item{start.year, end.year}{Starting and ending year for statistics in YYYY format, e.g. start.year=2013.}

\item{use.water.year}{Should results be computed on the water year (starting 1 October)?
The 2013/2014 water year
runs from 2013-10-01 to 2014-09-30.}

\item{roll.avg.days}{Volume frequency analysis conducted on these rolling averages.}

\item{use.log}{Transfrom to log-scale before analysis?}

\item{use.max}{Analyze the maximums rather than the minimums.}

\item{prob.plot.position}{Which plotting positions should be used in the frequency plots. Points are plotted
against  (i-a)/(n+1-a-b) where \code{i} is the rank of the value; \code{n} is the sample size and
\code{a} and \code{b} are defined as:
(a=0, b=0) for Weibull plotting positions;
(a=.2; b=.3) for Median plotting postions;
(a=.5; b=.5) for Hazen plotting positions.}

\item{prob.scale.points}{What points should be plotted along the $X$ axis in the frequency plot.}

\item{fit.distr}{Which distribution should be fit? PIII = Pearson Log III distribution; weibull=Weibull distribution.}

\item{fit.distr.method}{Which method used to fit the distribution. MOM=Method of moments; MLE=maximum likelihood estimation.}

\item{fit.quantiles}{Which quantiles should be estimated from the fitted distribution?}

\item{na.rm}{List of logical values indicating how to deal with missing values.
\code{na.rm$na.rm.global} control if missing values are to be globally excluded (TRUE) or propogated (FALSE).
If missing values are to be propogated, then any statistic computed over a missing value is also set
to missing. For example, if a flow value was missing on 2014-01-03, then any statistics for January of
2014, JFM of 2014, calendar year 2014, water year 2014, 3 day rolling average of 1, 2, 3 of January 2014
and all statistics that depend on those rolling averages, will also be set to missing.}

\item{write.stat.csv}{Should a file be created with the computed percentiles?
The file name will be  \code{file.path(report.dir,paste(Station.Code,"-annual-vfa-stat.csv", sep=""))}.}

\item{write.stat.trans.csv}{Should a file be created with the computed percentiles in transposed format?
The file name will be  \code{file.path(report.dir,paste(Station.Code,"-annual-vfa-stat-trans.csv", sep=""))}.}

\item{write.plotdata.csv}{Should a file be created with the frequency plot data?
The file name will be  \code{file.path(report.dir, paste(Station.Code,"-annual-vfa-plotdata.csv", sep=""))}.}

\item{write.quantiles.csv}{Should a file be created with the fitted quantiles?.
The file name will be  \code{file.path(report.dir, paste(Station.Code,"-annual-vfa-quantiles.csv", sep=""))}.}

\item{write.quantiles.trans.csv}{Should a file be created with the (transposed) fitted quantiles?
The file name will be \code{file.path(report.dir, paste(Station.Code,"-annual-vfa-quantiles-trans.csv", sep=""))}.}

\item{write.frequency.plot}{Should a file be created with the frequency plot..
The file name will be \code{file.path(report.dir, paste(Station.Code,"-annual-vfa-frequency-plot.",write.frequency.plot.suffix[1],sep=""))}}

\item{write.frequency.plot.suffix}{Format of the frequency plot.}

\item{report.dir}{Directory in which *.csv or *.pdf files are stored.}

\item{csv.nddigits}{Number of decimal digits to be rounded to when creating *.csv files.}

\item{debug}{Internal flag used for debugging}
}
\value{
A list with the following elements:
  \item{Q.flow.summary}{Data frame with flow summary.}
  \item{start.year}{Start year of the analysis}
  \item{end.year}{End year of the analysis}
  \item{use.water.year}{Were computations done on water year?}
  \item{use.max}{Were computations done on maximum values.}
  \item{roll.avg.days}{Rolling average days on which statistics computed.}
  \item{Q.stat}{Data frame with Computed annual summary statistics used in analysis}
  \item{Q.stat.trans}{Data frame with Computed annual summary statistics (transposed) used in analysis.}
  \item{plotdata}{Data frame with Co-ordinates used in frequency plot.}
  \item{prob.plot.position}{Which plotting position was used in frequency plot?}
  \item{freqplot}{ggplot2 object with frequency plot}
  \item{fit.distr}{Which distribution was fit to the data}
  \item{fit}{List of fitted objects from fitdistrplus.}
  \item{fitted.quantiles}{Data frame with fitted quantiles.}
  \item{fitted.quantiles.trans}{Data frane with fitted quantiles (transposed)}
  \item{file.stat.csv}{File name of *.csv file with flow summary}
  \item{ file.stat.trans.csv}{File name of *.csv file with flow summary (transposed)}
  \item{file.plotdata.csv}{File name of *.csv file with plotting information for frequency plot.}
  \item{file.quantile.csv}{File name of fitted quantiles.}
  \item{file.quantile.trans.csv}{File name of fitted quantiles (transposed)}
  \item{file.frequency.plot}{File name of *.pdf or *.png file with frequency plot}
  \item{Version}{Version of this function.}
  \item{Date}{Date function was run.}
}
\description{
Performs a volume frequency analysis on annual statistics similar to HEC-SSP.
}
\examples{
\dontrun{
vfa.analysis <- compute.volume.frequency.analysis(
                     Station.Code ='XXX',
                     flow         =flow,
                     start.year   =1960,
                     end.year     =2014)
}
}
