% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WSC.08HA011.R
\docType{data}
\name{WSC.08HA011}
\alias{WSC.08HA011}
\title{Stream discharge data from the Upper Cowichan River (Station WSC.08HA011).}
\format{A data frame with many rows and 2 variables.:
\describe{
  \item{Date}{Date as an R date class variable}
  \item{Q}{Average daily stream flow}  ...
}}
\description{
A dataset the date and daily average stream flow.
The basin area is 826 km**2.
}
\details{
While the file has dates ranging from 1900 onwards, data is only available
from 1965 onwards.Consequently, when analyzing this data you likely
want to set start.year=1965 and end.year=2015
}
\examples{
data(WSC.08HA011)
head(WSC.08HA011)
}
